chore(*): renamed httpException.ts to exception.ts
